requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.3
asgiref==3.3.1
Django==3.1.6
django-bootstrap-modal-forms==2.0.1
mysql==0.0.2
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
python-dotenv